requirements.txt
streamlit
pandas
numpy
openai
pinecone-client
langchain
langchain-community
python-dotenv